backend/requirements.txt
# FastAPI y servidor - BÁSICO PARA EMPEZAR
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Variables de entorno
python-dotenv==1.0.1

# Base de datos (añadiremos después)
# sqlalchemy==2.0.25
# mysql-connector-python==8.3.0

# Autenticación (añadiremos después cuando sea necesario)
# python-jose==3.3.0
# passlib[bcrypt]==1.7.4